Initial commit of BM 1.8

+++ src/main/java/WayofTime/alchemicalWizardry/common/spell/complex/effect/cse/wind/CSEProjectileDefensiveWind.java
@@ -40,6 +40,6 @@ public class CSEProjectileDefensiveWind extends ComplexSpellEffect
 	@Override
 	public int getCostOfEffect() 
 	{
-        return (100 * (this.potencyEnhancement + 1));
+        return (int) (100 * (this.potencyEnhancement + 1));
 	}
 }
